Refactor packages to mach with convention
