+++ src/managers/WebManager.java
@@ -1,4 +1,4 @@
-package Files;
+package managers;
 
 import java.io.BufferedReader;
 import java.io.IOException;
